Renamed events to future Events
Outsourced event preparation to utils.server
